Move linter code into top-level package
