+++ linter/src/main/java/com/facebook/coresql/linter/warning/StandardWarningCode.java
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-package com.facebook.coresql.warning;
+package com.facebook.coresql.linter.warning;
 
 public enum StandardWarningCode
 {
