+++ linter/src/main/java/com/facebook/coresql/linter/warning/WarningCollectorConfig.java
@@ -12,17 +12,23 @@
  * limitations under the License.
  */
 
-package com.facebook.coresql.warning;
+package com.facebook.coresql.linter.warning;
 
-import com.facebook.coresql.parser.AstNode;
+import static com.google.common.base.Preconditions.checkArgument;
 
-import java.util.List;
-
-public interface WarningCollector
+public class WarningCollectorConfig
 {
-    void add(WarningCode code, String warningMessage, AstNode node);
+    private int maxWarnings = Integer.MAX_VALUE;
 
-    List<CoreSqlWarning> getAllWarnings();
+    public WarningCollectorConfig setMaxWarnings(int maxWarnings)
+    {
+        checkArgument(maxWarnings >= 0, "maxWarnings must be >= 0");
+        this.maxWarnings = maxWarnings;
+        return this;
+    }
 
-    void clearWarnings();
+    public int getMaxWarnings()
+    {
+        return maxWarnings;
+    }
 }
