+++ pom.xml
@@ -1,22 +1,31 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
+
     <parent>
         <groupId>com.facebook.airlift</groupId>
         <artifactId>airbase</artifactId>
         <version>99</version>
     </parent>
+
     <groupId>com.facebook.presto</groupId>
     <artifactId>presto-coresql</artifactId>
     <version>0.2-SNAPSHOT</version>
     <packaging>pom</packaging>
+
     <name>presto-coresql</name>
+    <description>coreSql</description>
+    <url>https://github.com/prestodb/sql</url>
+
     <properties>
         <air.main.basedir>${project.parent.basedir}</air.main.basedir>
     </properties>
+
     <modules>
         <module>parser</module>
+        <module>linter</module>
     </modules>
+
     <build>
         <plugins>
             <plugin>
@@ -25,6 +34,7 @@
                 <version>0.3</version>
                 <extensions>true</extensions>
             </plugin>
+
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-release-plugin</artifactId>
@@ -32,6 +42,7 @@
                     <preparationGoals>clean verify -DskipTests</preparationGoals>
                 </configuration>
             </plugin>
+
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
@@ -39,6 +50,7 @@
                     <fork>false</fork>
                 </configuration>
             </plugin>
+
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-checkstyle-plugin</artifactId>
@@ -46,6 +58,7 @@
                     <excludes>**/generated-sources/**/*</excludes>
                 </configuration>
             </plugin>
+
             <plugin>
                 <!-- <groupId>org.apache.maven.plugins</groupId> -->
                 <artifactId>maven-pmd-plugin</artifactId>
@@ -53,6 +66,7 @@
                     <skip>true</skip>
                 </configuration>
             </plugin>
+
             <plugin>
                 <groupId>com.github.spotbugs</groupId>
                 <artifactId>spotbugs-maven-plugin</artifactId>
@@ -65,6 +79,7 @@
                     </execution>
                 </executions>
             </plugin>
+
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-checkstyle-plugin</artifactId>
@@ -72,6 +87,7 @@
                     <excludes>**/generated-sources/**/*</excludes>
                 </configuration>
             </plugin>
+
             <!-- Always build a jar with the test classes -->
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
